# application/services/custom_services/personmanager/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: healthmonkstart61 open api specification
   description: personmanager API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8012
     description: personmanager
tags:
- name: Person
  description: personal information such as PII
paths:
   "/Person/{id}":
      delete:
         tags:
         -  Person
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Person/get/search":
      get:
         tags:
         -  Person
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Person/get/update":
      put:
         tags:
         -  Person
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Person"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Person":
      put:
         tags:
         -  Person
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Person"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  Person
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  Person
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Person"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      Person:
         type: object
         properties:
            dob:
               type: "number"
               description: Description
            age:
               type: "number"
               description: Description
            gender-bio:
               type: "string"
               description: Description
            gender-preference:
               type: "string"
               description: Description
            metricsystem:
               type: "boolean"
               description: Description
            height:
               type: "string"
               description: Description
            currentweight:
               type: "number"
               description: Description
            currentactivitylevel:
               type: "string"
               description: Description
            foodpreference:
               type: "string"
               description: Description
            currentfitnesslevel:
               type: "string"
               description: Description
            medicalconditions:
               type: "string"
               description: Description
            otherconditions:
               type: "string"
               description: Description